fixes a bunch of unrelated issues (#4329)
* 1

* eye

* more

* maybe

* fix

* fix

+++ code/game/objects/items/weapons/melee/misc.dm
@@ -117,8 +117,9 @@
 
 /obj/item/melee/skateboard/dropped(mob/user, flags, atom/newLoc)
 	. = ..()
-	var/turf/T = get_turf(src)
-	new board_item_type(T)
+	if(!isturf(newLoc))
+		return
+	new board_item_type(newLoc)
 	qdel(src)
 
 /obj/item/melee/skateboard/improv
